chore(SRPRefactored, SRPViolated): move package
