+++ app/src/test/java/com/example/learningtest/solid/srp/SRPRefactored.kt
@@ -1,0 +1,42 @@
+package com.example.learningtest.solid.srp
+
+class SRPRefactored {
+    class LottoSeller {
+        fun soldLotto(money: Int): List<Lottery> {
+            val count = money / LOTTO_PRICE
+            return List(count) { LotteryGenerateStrategy().lotto() }
+        }
+
+        companion object {
+            private const val LOTTO_PRICE = 1_000
+        }
+    }
+
+    data class Lottery(val numbers: List<Int>) {
+        init {
+            numbers.forEach {
+                require(numbers.size == NUMBER_COUNT) {
+                    "Invalid lotto number count"
+                }
+                require(it in MIN_NUMBER..MAX_NUMBER) {
+                    "Invalid lotto number"
+                }
+            }
+        }
+
+        companion object {
+            private const val MIN_NUMBER = 1
+            private const val MAX_NUMBER = 45
+            val numberRange = (MIN_NUMBER..MAX_NUMBER)
+
+            const val NUMBER_COUNT = 6
+        }
+    }
+
+    class LotteryGenerateStrategy {
+        fun lotto(): Lottery =
+            Lottery(
+                (Lottery.numberRange).shuffled().take(Lottery.NUMBER_COUNT).sorted(),
+            )
+    }
+}
